Fixing issue from manual testing. Moving download handler to bzz file
